add generate file
